+++ app/src/main/java/com/huawei/facemask/GmsVision.java
@@ -25,7 +25,7 @@ class GmsVision {
     private static final Paint sFaceLandmarkPaint = new Paint();
     private static Detector<Face> sFaceDetector;
     private static long sRecognitionTime;
-    private static long sRecognitionTimeSum;
+    private static float sRecognitionTimeSum;
     private static long sRecognizedFrameCount;
     private static long sFrameCount;
 
@@ -89,10 +89,10 @@ class GmsVision {
                         @SuppressLint("StringFormatMatches")
                         @Override
                         public void run() {
-                            long percent = 0;
-                            long average = 0;
+                            float percent = 0;
+                            float average = 0;
                             if (sFrameCount != 0) {
-                                percent = sRecognizedFrameCount * 100 / sFrameCount;
+                                percent = (float) sRecognizedFrameCount * 100 / sFrameCount;
                                 average = sRecognitionTimeSum / sFrameCount;
                             }
                             score.setText(activity.getResources().getString(
