+++ app/src/main/java/com/huawei/facemask/HaarCascade.java
@@ -24,7 +24,7 @@ class HaarCascade {
     @SuppressWarnings({"FieldCanBeLocal", "unused"})
     private static boolean sDetectLandmarks;
     private static long sRecognitionTime;
-    private static long sRecognitionTimeSum;
+    private static float sRecognitionTimeSum;
     private static long sRecognizedFrameCount;
     private static long sFrameCount;
 
@@ -74,10 +74,10 @@ class HaarCascade {
                             @SuppressLint("StringFormatMatches")
                             @Override
                             public void run() {
-                                long percent = 0;
-                                long average = 0;
+                                float percent = 0;
+                                float average = 0;
                                 if (sFrameCount != 0) {
-                                    percent = sRecognizedFrameCount * 100 / sFrameCount;
+                                    percent = (float) sRecognizedFrameCount * 100 / sFrameCount;
                                     average = sRecognitionTimeSum / sFrameCount;
                                 }
                                 score.setText(activity.getResources().getString(
